+++ src/ResourceManager/Network/Commands.Network.Test/ScenarioTests/AzureFirewallFqdnTagTests.cs
@@ -31,7 +31,7 @@ namespace Commands.Network.Test.ScenarioTests
 
         [Fact]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
-        [Trait(Category.Owner, "azurefirewall")]
+        [Trait(Category.Owner, NrpTeamAlias.azurefirewall)]
         public void TestListAzureFirewallFqdnTag()
         {
             NetworkResourcesController.NewInstance.RunPsTest(_logger, "Test-ListAzureFirewallFqdnTag");