+++ src/ResourceManager/Network/Commands.Network/ChangeLog.md
@@ -18,11 +18,15 @@
         - Additional information about change #1
 --->
 ## Current Release
+
+## Version 6.9.1
 * Update cmdlet Test-AzureRmNetworkWatcherConnectivity, pass the protocol value to backend.
+* Added ResourceName argument completer to all cmdlets.
 * Added WhatIf support to Set-AzureRmNetworkSecurityGroup
 * Added ArgumentCompeter for Subnet's parameter ServiceEndpoints
 * Added cmdlet Get-AzureRmFirewallFqdnTag
 * Added Exclusion list and Global config functionality for application gateway WAF, new cmdlets added
+* Fixed issue with RouteFilter not being set if it was passed to ExpressRouteCircuit Peering as resource
 * Added ICMP back to supported AzureFirewall Network Protocols
 
 ## Version 6.9.0